Replaced pipeline runastrodriz with what was runaligndriz
